Renames the class
